requirements.txt
asn1crypto==0.23.0
awsebcli==3.11.0
blessed==1.14.2
boto==2.48.0
boto3==1.4.4
botocore==1.5.95
cement==2.8.2
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
colorama==0.3.7
cryptography==2.0.3
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-storages==1.6.1
djangorestframework==3.6.4
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
oauthlib==2.0.3
olefile==0.44
pathspec==0.5.0
Pillow==4.2.1
psycopg2==2.7.3.1
pycparser==2.18
PyJWT==1.5.3
pystache==0.5.4
python-dateutil==2.6.1
python-decouple==3.0
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
s3transfer==0.1.11
semantic-version==2.5.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
static==1.1.1
static3==0.7.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.44.0
whitenoise==3.3.1
